+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Parse VELUX ACTIVE indoor climate sensors (`NXS`), departure switches (`NXD`),
+  and room climate data.
+
 ## [9.6.0] - 2026-07-28
 
 ### Added